Use the unoptimized flow
